+++ lib/app_theme.dart
@@ -1,0 +1,4 @@
+library app_theme;
+
+export 'theme/material_theme.dart';
+export 'theme/text_theme.dart';