+++ lib/bottom_sheets/bottom_sheets.dart
@@ -1,3 +1,0 @@
-library bottom_sheets;
-
-export 'input_bottom_sheet.dart';