+++ lib/dialogs.dart
@@ -1,0 +1,3 @@
+library dialogs;
+
+export 'dialogs/confirmation_dialog.dart';