+++ lib/dialogs/dialogs.dart
@@ -1,3 +1,0 @@
-library dialogs;
-
-export 'confirmation_dialog.dart';