+++ lib/extensions.dart
@@ -1,0 +1,5 @@
+library extensions;
+
+export 'extensions/list_extension.dart';
+export 'extensions/string_extension.dart';
+export 'extensions/widget_extension.dart';