+++ lib/extensions/extensions.dart
@@ -1,5 +1,0 @@
-library extensions;
-
-export 'list_extension.dart';
-export 'string_extension.dart';
-export 'widget_extension.dart';