+++ lib/theme/app_theme.dart
@@ -1,4 +1,0 @@
-library app_theme;
-
-export 'material_theme.dart';
-export 'text_theme.dart';