+++ lib/widgets.dart
@@ -1,0 +1,4 @@
+library widgets;
+
+export 'widgets/input_box.dart';
+export 'widgets/message.dart';