+++ lib/widgets/message.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:utils/extensions/extensions.dart';
-import 'package:utils/theme/app_theme.dart' as AppTheme;
+import 'package:utils/app_theme.dart' as AppTheme;
+import 'package:utils/extensions.dart';
 
 /// Displays a widget and message centered on a page
 class Message extends StatelessWidget {