+++ lib/widgets/widgets.dart
@@ -1,4 +1,0 @@
-library widgets;
-
-export 'input_box.dart';
-export 'message.dart';