+++ test/list_extension_intersect_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import 'package:utils/extensions/extensions.dart';
+import 'package:utils/extensions.dart';
 
 void main() {
   test("intersect: list of integers -> [1, 2, 3] ∩ [2, 3] -> [2, 3]", () {