+++ test/list_extension_remove_blank_items_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import 'package:utils/extensions/extensions.dart';
+import 'package:utils/extensions.dart';
 
 void main() {
   test("removeBlankItems: empty list of strings", () {