+++ test/string_capitalize_test.dart
@@ -1,67 +1,23 @@
 import 'package:test/test.dart';
-import 'package:utils/extensions/string_extension.dart';
+import 'package:utils/extensions.dart';
 
 void main() {
-  test("all lower case: test this beautiful text", () {
-    const String text = "test this beautiful text";
-
+  test("Capitalize text -> generic cases", () {
     const String expected = "Test this beautiful text";
-    final String actual = text.capitalize();
 
-    expect(actual, expected);
+    expect("test this beautiful text".capitalize(), expected);
+    expect("Test this beautiful text".capitalize(), expected);
+    expect("TEST THIS BEAUTIFUL TEXT".capitalize(), expected);
+    expect("Test This Beautiful Text".capitalize(), expected);
+    expect("TeSt this BeAUTifUL TExT".capitalize(), expected);
   });
 
-  test("all upper case: TEST THIS BEAUTIFUL TEXT", () {
-    const String text = "TEST THIS BEAUTIFUL TEXT";
+  test("Capitalize text -> special cases", () {
+    final String expected = StringExtension.empty();
 
-    const String expected = "Test this beautiful text";
-    final String actual = text.capitalize();
-
-    expect(actual, expected);
-  });
-
-  test("first word upper case: Test this beautiful text", () {
-    const String text = "Test this beautiful text";
-
-    const String expected = "Test this beautiful text";
-    final String actual = text.capitalize();
-
-    expect(actual, expected);
-  });
-
-  test("every word upper case: Test This Beautiful Text", () {
-    const String text = "Test This Beautiful Text";
-
-    const String expected = "Test this beautiful text";
-    final String actual = text.capitalize();
-
-    expect(actual, expected);
-  });
-
-  test("mixed case: TeSt this BeAUTifUL TExT", () {
-    const String text = "TeSt this BeAUTifUL TExT";
-
-    const String expected = "Test this beautiful text";
-    final String actual = text.capitalize();
-
-    expect(actual, expected);
-  });
-
-  test("White spaces", () {
-    const String text = "  ";
-
-    const String expected = "";
-    final String actual = text.capitalize();
-
-    expect(actual, expected);
-  });
-
-  test("Empty string", () {
-    const String text = "";
-
-    const String expected = "";
-    final String actual = text.capitalize();
-
-    expect(actual, expected);
+    expect("  ".capitalize(), expected);
+    expect("".capitalize(), expected);
+    expect('  '.capitalize(), expected);
+    expect(''.capitalize(), expected);
   });
 }