+++ test/string_empty_test.dart
@@ -1,18 +1,11 @@
 import 'package:test/test.dart';
-import 'package:utils/extensions/extensions.dart';
+import 'package:utils/extensions.dart';
 
 void main() {
-  test("empty: string is empty -> single quot", () {
-    const String expected = '';
+  test("empty: string is empty", () {
     final String actual = StringExtension.empty();
 
-    expect(actual, expected);
-  });
-
-  test("empty: string is empty -> double quote", () {
-    const String expected = "";
-    final String actual = StringExtension.empty();
-
-    expect(actual, expected);
+    expect(actual, '');
+    expect(actual, "");
   });
 }